backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
pydantic-settings==2.5.0
python-dotenv==1.0.1
requests==2.31.0
httpx==0.27.0
langgraph==0.2.28
langchain-core==0.3.0

